% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{lmms-class}
\alias{lmms-class}
\title{\code{lmms} class a S4 superclass to extend \code{lmmspline}  and \code{lmmsde} class.}
\description{
\code{lmms} class is a superclass for classes \code{lmmspline}  and \code{lmmsde}. These classes inherit common slots.
}
\section{Slots}{

\describe{
\item{\code{basis}}{An object of class \code{character} describing the basis used for modelling.}

\item{\code{knots}}{An object of class \code{numeric}, describing the boundaries of the splines.}

\item{\code{errorMolecules}}{Vector of class \code{character}, describing the molecules that could not be modelled.}
}}

