% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{lmmsde-class}
\alias{lmmsde-class}
\title{\code{lmmsde} class a S4 class that extends \code{lmms} class.}
\description{
\code{lmmsde} class inherits from class \code{lmms} and extends it with six further slots: DE, model.time, model.group, model.time.group, type and experiment. The class \code{lmmsde} is returned when applying \code{\link{lmmsDE}} method.
}
\section{Slots}{

\describe{
\item{\code{DE}}{A \code{data.frame} returning p-values and adjusted p-values using Benjamini-Hochberg correction for multiple testing of the differential expression testing over time, group or their interaction.}

\item{\code{modelTime}}{A \code{list} of class\code{\link{lme}}, containing the models for every molecule modelling the time effect.}

\item{\code{modelGroup}}{A \code{list} of class \code{\link{lme}}, containing the models for every molecule modelling group effect.}

\item{\code{modelTimeGroup}}{A \code{list} of class \code{\link{lme}}, containing the models for every molecule modelling time and group interaction effect.}

\item{\code{type}}{An object of class \code{character}, describing the test performed.}

\item{\code{experiment}}{An object of class \code{character} describing the model used to perform differential expression analysis.}
}}

