% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{lmmspline-class}
\alias{lmmspline-class}
\title{\code{lmmspline} class a S4 class that extends \code{lmms} class.}
\description{
\code{lmmspline} class inherits from class \code{lmms} and extends it with three further slots: \code{predSpline}, \code{modelsUsed}, \code{models}. The class is returned when applying \code{\link{lmmSpline}} method.
}
\section{Slots}{

\describe{
\item{\code{predSpline}}{A \code{data.frame} returning the fitted values for the time points of interest.}

\item{\code{modelsUsed}}{A vector of class \code{numeric}, describing the polynomial boundaries of the splines. If not defined or if basis='cubic' knots are automatically estimated using Ruppert 2002 or are the design points when using 'cubic'.}

\item{\code{models}}{A \code{list} of class \code{lm} or \code{lme}, containing the models used to model the particular feature of interest.}

\item{\code{data}}{A \code{data.frame} containing the data used to model the particular molecule of interest.}

\item{\code{derivative}}{A \code{logical} value indicating if the derivative was calculated.}
}}

