% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{noise-class}
\alias{noise-class}
\title{\code{noise} S4 class}
\description{
The class \code{noise} is returned when applying \code{\link{investNoise}} method.
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character} vector. The name of the trajectory.}

\item{\code{RT}}{A \code{numeric} vector, containing the time to molecule standard deviation ratios for every trajectory.}

\item{\code{RI}}{A \code{numeric} vector, containing the individual to molecule standard deviation ratios for every trajectory.}

\item{\code{propMissing}}{A \code{numeric} vector, containing the proportion of missing values for every trajectory.}

\item{\code{foldChange}}{A \code{numeric} vector, containing the maximum fold change of the mean between any two time points.}
}}

